Moving Graphics*CreateInfo to the TerraFX.Advanced namespace
